requirements.txt
Django==4.0.3
django_redis==5.2.0
django-cors-headers==3.13
djangorestframework==3.14
djoser==2.1
openpyxl==3.0.10
djangorestframework-recursive==0.1.2
Pillow==9.2.0
psycopg2-binary==2.9.4
gunicorn==20.1